Changed vect2

+++ src/com/skanderj/gingerbread3/math/Vector2.java
@@ -8,13 +8,13 @@ import com.skanderj.gingerbread3.util.Utilities;
  *
  */
 public class Vector2 {
-	public int x, y;
+	public float x, y;
 
 	public static final Vector2[] randomVectors(final int count, final int xMin, final int xMax, final int yMin, final int yMax) {
 		final Vector2[] vects = new Vector2[count];
 		for (int i = 0; i < count; i += 1) {
-			final int randomX = Utilities.randomInteger(xMin, xMax);
-			final int randomY = Utilities.randomInteger(yMin, yMax);
+			final float randomX = Utilities.randomInteger(xMin, xMax);
+			final float randomY = Utilities.randomInteger(yMin, yMax);
 			vects[i] = new Vector2(randomX, randomY);
 		}
 		return vects;
@@ -23,7 +23,7 @@ public class Vector2 {
 	/**
 	 * Self explanatory.
 	 */
-	public Vector2(final int x, final int y) {
+	public Vector2(final float x, final float y) {
 		this.x = x;
 		this.y = y;
 	}
